Avoid loading common.annex unless needed.
