+++ datalad_service/handlers/dataset.py
@@ -1,6 +1,6 @@
 import falcon
 
-from datalad_service.common.annex import get_user_info
+from datalad_service.common.user import get_user_info
 from datalad_service.common.celery import dataset_queue
 from datalad_service.tasks.dataset import *
